admin_server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiosqlite==0.20.0
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
pydantic>=2.7.0
httpx==0.27.0
